Decimal letter spacing

+++ src/pagetypes/MainPage.cpp
@@ -72,7 +72,7 @@ void MainPage::AddName()
     doc.AddText(config.getString("name"),
                 MIDDLE_LINE / 2 - 72,
                 14,
-                UtilType::TextOptions(FontType::Seagram, 34, 0, config.getInt("nameLetterSpacing")));
+                UtilType::TextOptions(FontType::Seagram, 34, 0, config.getDouble("nameLetterSpacing")));
 }
 
 void MainPage::AddRace()
